Unify ECB classes in AbstractECBCurrentRateProvider.java
