+++ src/main/java/org/javamoney/moneta/convert/internal/IMFRateProvider.java
@@ -129,7 +129,8 @@ public class IMFRateProvider extends AbstractRateProvider implements LoaderListe
         }
     }
 
-    private void loadRatesTSV(InputStream inputStream) throws IOException, ParseException {
+    @SuppressWarnings("unchecked")
+	private void loadRatesTSV(InputStream inputStream) throws IOException, ParseException {
         Map<CurrencyUnit, List<ExchangeRate>> newCurrencyToSdr = new HashMap<>();
         Map<CurrencyUnit, List<ExchangeRate>> newSdrToCurrency = new HashMap<>();
         NumberFormat f = new DecimalFormat("#0.0000000000");
